feat(budgapp): Add new page for web storage options
