+++ src/budgapp/src/client/general/GeneralView.scss
@@ -31,15 +31,4 @@
             }
         }
     }
-
-    .import-excel {
-        input[type="file"] {
-            display: none;
-        }
-
-        border: 1px solid black;
-        display: inline-block;
-        padding: 2px 4px;
-        font-size: 14px;
-    }
 }
